Renamed namespace Snake.Game to just Snake
