+++ src/Snake/Commands/KeyboardCommandReader.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Doog.Framework;
 
-namespace Snake.Game.Commands
+namespace Snake.Commands
 {
     public class KeyboardCommandReader : ICommandReader
     {
