+++ src/Snake/Scenes/Samples/Sample7Scene.cs
@@ -1,6 +1,6 @@
 ﻿using Doog.Framework;
 
-namespace Snake.Game.Scenes.Samples
+namespace Snake.Scenes.Samples
 {
     public class Sample7Scene : SceneBase
     {
