+++ src/Snake/Scenes/Samples/SampleComponent.cs
@@ -1,6 +1,6 @@
 ﻿using Doog.Framework;
 
-namespace Snake.Game.Scenes.Samples
+namespace Snake.Scenes.Samples
 {
     public class SampleComponent : RectangleComponent, IDrawable
     {
