+++ src/Snake/Score.cs
@@ -1,6 +1,6 @@
 ﻿using Doog.Framework;
 
-namespace Snake.Game
+namespace Snake
 {
     public class Score : RectangleComponent, IDrawable
     {
