+++ src/Snake/Snake.csproj
@@ -42,35 +42,35 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Game\Commands\KeyBinding.cs" />
-    <Compile Include="Game\Commands\KeyboardCommandReader.cs" />
-    <Compile Include="Game\Commands\MoveDownCommand.cs" />
-    <Compile Include="Game\Commands\MoveLeftCommand.cs" />
-    <Compile Include="Game\Commands\MoveRightCommand.cs" />
-    <Compile Include="Game\Commands\MoveUpCommand.cs" />
-    <Compile Include="Game\Commands\SnakeCommandPool.cs" />
-    <Compile Include="Game\FoodSpawner.cs" />
-    <Compile Include="Game\Food.cs" />
-    <Compile Include="Game\Snake.cs" />
-    <Compile Include="Game\SnakeGame.cs" />
-    <Compile Include="Game\SnakeTile.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Game\Wall.cs" />
-    <Compile Include="Game\WallSpawner.cs" />
-    <Compile Include="Game\Scenes\ClassicModeLevelScene.cs" />
-    <Compile Include="Game\Scenes\GameOverScene.cs" />
-    <Compile Include="Game\Scenes\Samples\Sample1Scene.cs" />
-    <Compile Include="Game\Scenes\Samples\Sample2Scene.cs" />
-    <Compile Include="Game\Scenes\Samples\Sample3Scene.cs" />
-    <Compile Include="Game\Scenes\Samples\SampleComponent.cs" />
-    <Compile Include="Game\Scenes\Samples\Sample4Scene.cs" />
-    <Compile Include="Game\Scenes\Samples\Sample5Scene.cs" />
-    <Compile Include="Game\Score.cs" />
-    <Compile Include="Game\Scenes\Samples\Sample6Scene.cs" />
-    <Compile Include="Game\Portal.cs" />
-    <Compile Include="Game\PortalBridge.cs" />
-    <Compile Include="Game\Scenes\Samples\Sample7Scene.cs" />
-    <Compile Include="Game\Scenes\Samples\Sample8Scene.cs" />
+    <Compile Include="Commands\KeyBinding.cs" />
+    <Compile Include="Commands\KeyboardCommandReader.cs" />
+    <Compile Include="Commands\MoveDownCommand.cs" />
+    <Compile Include="Commands\MoveLeftCommand.cs" />
+    <Compile Include="Commands\MoveRightCommand.cs" />
+    <Compile Include="Commands\MoveUpCommand.cs" />
+    <Compile Include="Commands\SnakeCommandPool.cs" />
+    <Compile Include="Scenes\ClassicModeLevelScene.cs" />
+    <Compile Include="Scenes\GameOverScene.cs" />
+    <Compile Include="Scenes\Samples\Sample1Scene.cs" />
+    <Compile Include="Scenes\Samples\Sample2Scene.cs" />
+    <Compile Include="Scenes\Samples\Sample3Scene.cs" />
+    <Compile Include="Scenes\Samples\Sample4Scene.cs" />
+    <Compile Include="Scenes\Samples\Sample5Scene.cs" />
+    <Compile Include="Scenes\Samples\Sample6Scene.cs" />
+    <Compile Include="Scenes\Samples\Sample7Scene.cs" />
+    <Compile Include="Scenes\Samples\Sample8Scene.cs" />
+    <Compile Include="Scenes\Samples\SampleComponent.cs" />
+    <Compile Include="Food.cs" />
+    <Compile Include="FoodSpawner.cs" />
+    <Compile Include="Portal.cs" />
+    <Compile Include="PortalBridge.cs" />
+    <Compile Include="Score.cs" />
+    <Compile Include="Snake.cs" />
+    <Compile Include="SnakeGame.cs" />
+    <Compile Include="SnakeTile.cs" />
+    <Compile Include="Wall.cs" />
+    <Compile Include="WallSpawner.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
@@ -84,8 +84,8 @@
   <ItemGroup>
     <Folder Include="Resources\" />
     <Folder Include="Resources\Fonts\" />
-    <Folder Include="Game\Scenes\" />
-    <Folder Include="Game\Scenes\Samples\" />
+    <Folder Include="Commands\" />
+    <Folder Include="Scenes\" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="..\Doog.Framework\Texts\Map\Resources\Fonts\Default.txt">
