+++ src/Snake/Wall.cs
@@ -1,6 +1,6 @@
 ﻿using Doog.Framework;
 
-namespace Snake.Game
+namespace Snake
 {
     public class Wall : RectangleComponent, IDrawable, ICollidable
     {
